move `eager_resolve_vars` to `rustc_type_ir`
